common LCM object and message handling for all LCM provider in DART
All LCM related classes in DART should use a common LCM instance and handle
messages in a central place. The common LCM object is instantiated once when
the first inherited class is instantiated and is shared over all inherited
classes. With its instantiation, a thread that handles incoming messages is
started and exists over the lifetime of the process.